requirements.txt
asgiref==3.2.7
backcall==0.1.0
beautifulsoup4==4.9.1
colorama==0.4.3
decorator==4.4.2
Django==2.2.12
djangorestframework==3.11.0
google==2.0.3
grpcio==1.29.0
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.17.0
joblib==0.14.1
numpy==1.18.4
pandas==0.25.3
parso==0.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
protobuf==3.12.2
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2020.1
schedule==0.6.0
scikit-learn==0.22.2.post1
scipy==1.4.1
shap==0.35.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
threadpoolctl==2.1.0
tqdm==4.46.0
traitlets==4.3.3
wcwidth==0.1.9
win-unicode-console==0.5
